refactor:page.tsxが長すぎたから修正
コンポーネントごとにファイルを分けた
